Fix minor MSVC warnings

+++ src/core/http/server_poll.cc
@@ -495,7 +495,7 @@ http_IO *http_Dispatcher::InitClient(int sock, int64_t start, struct sockaddr *s
     http_IO *client = nullptr;
 
     if (free_clients.len) {
-        int idx = GetRandomInt(0, free_clients.len);
+        int idx = GetRandomInt(0, (int)free_clients.len);
 
         client = free_clients[idx];
